refactor: resolve #776 - split useComposerPlayback test file to stay under 500-line limit
Extract frequency calculation tests (4 tests) into separate file.
Original: 530 lines → 452 lines. New frequency file: 113 lines.
